requirements.txt
mysql-connector-python==8.0.18
speedtest-cli==2.1.2
schedule==0.6.0